use cargo to bootstrap worker
The regular rust_binary() and rust_library() rules require a worker
toolchain, even if it is a dummy one. This prevents us from using
these rules to build the worker binary, as it leads to a cyclic
dependency.

Distributing an external binary is not ideal - it's hard to verify
what it's doing, and awkward to add on new platforms. To work around
this, we invoke cargo directly, using the configured toolchain. This
causes cargo to fetch the dependencies directly, and
build the worker binary which we can then use for future compilation.